refactor. extract HttpResponse.Body class
HttpResponse 코드가 길어지는 문제로 Body 클래스 및 자식 클래스를 다른 패키지로 추출
